+++ lib/enums.dart
@@ -3,3 +3,5 @@ enum TextSize { normal, medium, label, title, subTitle, large, xl }
 enum TextColors { primary, secondary, accent, success, danger, info, warning }
 
 enum ButtonType { text, icon, iconText, textIcon }
+
+enum ToastType { success, error, info }